+++ src/main/java/me/gurkz/superdupermod/mixin/LeashableVillagerMixin.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2025 Gurkan
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package me.gurkz.superdupermod.mixin;
+
+import net.minecraft.entity.passive.MerchantEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(MerchantEntity.class)
+public class LeashableVillagerMixin {
+    @Inject(
+            method = "canBeLeashed",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    private void canBeLeashed(CallbackInfoReturnable<Boolean> cir) {
+        cir.setReturnValue(true);
+        cir.cancel();
+    }
+}